create_outlines

Adds vertical and horizontal bars to the table.

Usage

create_outlines(tbl, workbook, sheet, locations, styles)

Arguments

tbl

table created with tablespan::tablespan

workbook

Excel workbook created with openxlsx::createWorkbook()

sheet

name of the sheet to which the table should be written to

locations

list with overview of row and col locations for different table elements

styles

openxlsx style for the different table elements (see ?tablespan::tab_styles). The styles element also allows applying custom styles to parts of the data shown in the table body.