get_variables_from_list

Extracts the variable names from the left or right hand side of a deparsed formula (see ?tablespan:::deparse_formula).

Usage

get_variables_from_list(deparsed_formula_element, variables = c())

Arguments

deparsed_formula_element

left or right hand side of the result from tablespan:::deparse_formula

variables

the function is recursive and fills the variable vector

Value

a vector with the names of the variables

Examples

library(tablespan)
deparsed <- tablespan:::deparse_formula(formula =
 (`Row Name` = `Row 1` + `Row 2`) ~ `Column 1` + (`Column Banner` = `Column 2` + `Column 3`))
str(deparsed)
#> List of 2
#>  $ lhs:List of 2
#>   ..$ name   : chr "_BASE_LEVEL_"
#>   ..$ entries:List of 1
#>   .. ..$ :List of 2
#>   .. .. ..$ name   : chr "Row Name"
#>   .. .. ..$ entries:List of 2
#>   .. .. .. ..$ :List of 3
#>   .. .. .. .. ..$ name     : chr "Row 1"
#>   .. .. .. .. ..$ item_name: chr "Row 1"
#>   .. .. .. .. ..$ entries  : NULL
#>   .. .. .. ..$ :List of 3
#>   .. .. .. .. ..$ name     : chr "Row 2"
#>   .. .. .. .. ..$ item_name: chr "Row 2"
#>   .. .. .. .. ..$ entries  : NULL
#>  $ rhs:List of 2
#>   ..$ name   : chr "_BASE_LEVEL_"
#>   ..$ entries:List of 2
#>   .. ..$ :List of 3
#>   .. .. ..$ name     : chr "Column 1"
#>   .. .. ..$ item_name: chr "Column 1"
#>   .. .. ..$ entries  : NULL
#>   .. ..$ :List of 2
#>   .. .. ..$ name   : chr "Column Banner"
#>   .. .. ..$ entries:List of 2
#>   .. .. .. ..$ :List of 3
#>   .. .. .. .. ..$ name     : chr "Column 2"
#>   .. .. .. .. ..$ item_name: chr "Column 2"
#>   .. .. .. .. ..$ entries  : NULL
#>   .. .. .. ..$ :List of 3
#>   .. .. .. .. ..$ name     : chr "Column 3"
#>   .. .. .. .. ..$ item_name: chr "Column 3"
#>   .. .. .. .. ..$ entries  : NULL
tablespan:::get_variables_from_list(deparsed$lhs)
#> [1] "Row 1" "Row 2"