write_header_entry

Recursive function writing the header elements (column names and names for the rownames) to the workbook.

Usage

write_header_entry(
  workbook,
  sheet,
  header_entry,
  max_level,
  start_row,
  start_col,
  header_style,
  vline_style
)

Arguments

workbook

Excel workbook created with openxlsx::createWorkbook()

sheet

name of the sheet to which the table should be written to

header_entry

specific header enty to write to the workbook

max_level

the highest level of header entries

start_row

integer specifying row to write to

start_col

integer specifying column to write to

header_style

openxlsx style for the header

vline_style

openxlsx style for the vertical lines in the header