as_gt

Translates a table created with tablespan to a great table (gt). See <https://gt.rstudio.com/>.

Usage

as_gt(
  tbl,
  groupname_col = NULL,
  separator_style = gt::cell_borders(sides = c("right"), weight = gt::px(1), color =
    "gray"),
  auto_format = TRUE,
  ...
)

Arguments

tbl

table created with tablespan::tablespan

groupname_col

Provide column names to group data. See ?gt::gt for more details.

separator_style

style of the vertical line that separates the row names from the data.

auto_format

should the table be formatted automatically?

...

additional arguments passed to gt::gt().

Value

gt table that can be further adapted with the gt package.

Details

Tablespan itself does not provide any printing of tables as HTML table. However, with as_gt, tablespan can be translated to a great table which provides html and LaTeX output.

Examples

library(tablespan)
library(dplyr)
data("mtcars")

summarized_table <- mtcars |>
  group_by(cyl, vs) |>
  summarise(N = n(),
            mean_hp = mean(hp),
            sd_hp = sd(hp),
            mean_wt = mean(wt),
            sd_wt = sd(wt))
#> `summarise()` has grouped output by 'cyl'. You can override using the `.groups`
#> argument.

tbl <- tablespan(data = summarized_table,
                 formula = (LHS = Cylinder:cyl + Engine:vs) ~
                   N +
                   (Results = (`Horse Power` = Mean:mean_hp + SD:sd_hp) +
                      (`Weight` = Mean:mean_wt + SD:sd_wt)))

gt_tbl <- as_gt(tbl)
gt_tbl

  

      	
        
Results

      
    

        
LHS

      
      	N
      	
        
Horse Power

      
      	
        
Weight

      
    

Cylinder
      	Engine
      	Mean
      	SD
      	Mean
      	SD
    

4 
	0 
	 1 
	 91    

	2.14 

4 
	1 
	10 
	 81.8  
	21.872
	2.3  
	0.598

6 
	0 
	 3 
	131.667
	37.528
	2.755
	0.128

6 
	1 
	 4 
	115.25 
	 9.179
	3.389
	0.116

8 
	0 
	14 
	209.214
	50.977
	3.999
	0.759